refactor: 리렉토리 구조 변경
