+++ STEP2-REQUIREMENTS.md
@@ -7,12 +7,12 @@
 
 ### 필수 요구사항
 
-- [ ] Step1 미션 `나만의 Virtual DOM` 사용
-- [ ] `데이터 변경에 반응`하여 렌더링되도록 구현
+- [x] Step1 미션 `나만의 Virtual DOM` 사용
+- [x] `데이터 변경에 반응`하여 렌더링되도록 구현
 
 ### 선택 요구사항 (택 1)
 
-- [ ] 나만의 상태 관리 패턴 사용
+- [x] 나만의 상태 관리 패턴 사용
 - [ ] Proxy API 사용
 
 ### 심화 요구사항
